refactor: 메일 전송, 인증 요청 방식 변경
