+++ src/main/java/com/bob/web/auth/mail/request/MailSendRequest.java
@@ -1,0 +1,7 @@
+package com.bob.web.auth.mail.request;
+
+public record MailSendRequest(
+    String email
+) {
+
+}
