+++ src/main/java/com/bob/web/auth/mail/request/MailVerifyRequest.java
@@ -1,0 +1,8 @@
+package com.bob.web.auth.mail.request;
+
+public record MailVerifyRequest(
+    String email,
+    String code
+) {
+
+}
