+++ src/test/unit/java/com/bob/web/auth/mail/MailControllerTest.java
@@ -6,6 +6,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;
 
+import com.bob.domain.member.service.port.MailService;
+import com.bob.web.auth.mail.controller.MailController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,7 +16,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import com.bob.domain.member.service.port.MailService;
 
 @DisplayName("이메일 API 테스트")
 @ExtendWith(MockitoExtension.class)
